chore(coordinator): remove build prefix from package names
