+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/l1/LineaValidiumEnhancedWrapper.kt
@@ -1,6 +1,6 @@
 package net.consensys.linea.contract.l1
 
-import build.linea.contract.ValidiumV1
+import linea.contract.ValidiumV1
 import linea.web3j.transactionmanager.AsyncFriendlyTransactionManager
 import net.consensys.linea.contract.Web3JContractAsyncHelper
 import org.web3j.abi.datatypes.Function
