+++ coordinator/clients/smart-contract-client/src/main/kotlin/net/consensys/linea/contract/l1/Web3JLineaRollupFunctionBuilders.kt
@@ -1,6 +1,6 @@
 package net.consensys.linea.contract.l1
 
-import build.linea.contract.LineaRollupV6
+import linea.contract.LineaRollupV6
 import linea.contract.l1.LineaRollupContractVersion
 import linea.domain.BlobRecord
 import linea.domain.ProofToFinalize
