+++ jvm-libs/generic/extensions/tuweni/src/main/kotlin/linea/tuweni/Bytes32.kt
@@ -1,4 +1,4 @@
-package build.linea.tuweni
+package linea.tuweni
 
 import linea.kotlin.toULong
 import org.apache.tuweni.bytes.Bytes32
