+++ jvm-libs/generic/serialization/jackson/src/main/kotlin/linea/s11n/jackson/InstantISO8601SerDe.kt
@@ -1,4 +1,4 @@
-package build.linea.s11n.jackson
+package linea.s11n.jackson
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParser
@@ -8,14 +8,14 @@ import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
 import kotlin.time.Instant
 
-object InstantAsHexNumberSerializer : JsonSerializer<Instant>() {
+object InstantISO8601Serializer : JsonSerializer<Instant>() {
   override fun serialize(value: Instant, gen: JsonGenerator, serializers: SerializerProvider) {
-    gen.writeString("0x${value.epochSeconds.toString(16)}")
+    gen.writeString(value.toString())
   }
 }
 
-object InstantAsHexNumberDeserializer : JsonDeserializer<Instant>() {
+object InstantISO8601Deserializer : JsonDeserializer<Instant>() {
   override fun deserialize(p: JsonParser, ctxt: DeserializationContext): Instant {
-    return Instant.fromEpochSeconds(p.text.replace("0x", "").toLong(16))
+    return Instant.parse(p.text)
   }
 }
