+++ jvm-libs/generic/serialization/jackson/src/test/kotlin/linea/s11n/jackson/BytesSerDeTest.kt
@@ -1,4 +1,4 @@
-package build.linea.s11n.jackson
+package linea.s11n.jackson
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
